Move get_private_key endpoint to rita_common
