+++ rita_common/src/dashboard/mod.rs
@@ -3,9 +3,13 @@
 //! from the outside world for obvious security reasons.
 
 pub mod babel;
+pub mod backup_created;
+pub mod contact_info;
 pub mod debts;
 pub mod development;
+pub mod eth_private_key;
 pub mod interfaces;
+pub mod localization;
 pub mod nickname;
 pub mod own_info;
 pub mod settings;
@@ -15,6 +19,3 @@ pub mod usage;
 pub mod wallet;
 pub mod wg_key;
 pub mod wifi;
-pub mod contact_info;
-pub mod localization;
-pub mod backup_created;
